Changes related to new JSON format
1) Instead of scan created and updated date, added BOM updated date. In
case of import or deletion of code location, the scan created and
updated date won’t be available.
2) Instead of having project and version link at Project level, moving
it to component level. It is support the many-one mapping. In case of
components found across the mapped Hub projects, the project name will
be “Multiple Projects” and version name will be “Multiple Versions”
3) Instead of having href as the attribute name in the
VulnerabilityWithRemediationView, added Hub Vulnerability URL and NVD
Vulnerability URL.