Phone Home functionality and refactor code
**Description:**
Added a new job to send usage data to Black Duck. This job runs on the
same schedule as the push scan data task. In the process refactor code
so that PropertyConstants was no longer statically accessible (accessing
the static variables when the Spring bean was being instantiated cause
null return values and consequently validation methods were failing).

* Edit application.properties and application_prod.properties to have a
placeholder for the plugin version, and edit build.gradle to fill in
that placeholder using values from the project properties
* Ensure PropertyConstants is loaded as a Spring Bean
* Remove _static_ keyword from all PropertyConstants fields and methods
* Edit all locations where PropertyConstants is used to either provide a
Dependency Injected bean or pass the bean through the constructor for
non- spring bean classes
* Add new Job Configuration and Step classes for Phone Home capability.
Job runs on the same schedule as the push scan data task
* Modify all JUnit tests to use Dependency Injected beans where possible